DividesByZero (#13)
* DividesByZero

* Update MyNumber.java

To stay consistent

* fix: Removed files that should not be commited

* fix: Fixed Double to double to stay consistent with everything else

* fix: Fixed tests to work with double values instead of intergers

* fix: Fixed tests from the merge to use doubles

* Add the scenario and steps to fulfill test for the division by zero

* Add another test

* remove a test

* fix: Fixed Integer to Double from the merge + fixed division by zero tests

* fix: Forgot to add the cucumber tests for the new tests

* Modify integer to double in code

* fix : remove an uneccessary comments

* fix: Removed files that should not be commited + Modified .gitignore to ignore these files

---------

+++ src/main/java/calculator/Minus.java
@@ -25,12 +25,12 @@ public final class Minus extends Operation
   }
 
     /**
-     * The actual computation of the (binary) arithmetic subtraction of two integers
-     * @param l The first integer
-     * @param r The second integer that should be subtracted from the first
-     * @return The integer that is the result of the subtraction
+     * The actual computation of the (binary) arithmetic subtraction of two doubles
+     * @param l The first double
+     * @param r The second double that should be subtracted from the first
+     * @return The double that is the result of the subtraction
      */
-  public int op(int l, int r) {
+  public double op(double l, double r) {
   	return (l-r);
   }
 }
