+++ src/main/java/calculator/MyNumber.java
@@ -11,20 +11,20 @@ import visitor.Visitor;
  */
 public class MyNumber implements Expression
 {
-  private final int value;
+  private final double value;
 
     /** getter method to obtain the value contained in the object
      *
-     * @return The integer number contained in the object
+     * @return The double number contained in the object
      */
-  public Integer getValue() { return value; }
+  public double getValue() { return value; }
 
     /**
      * Constructor method
      *
-     * @param v The integer value to be contained in the object
+     * @param v The double value to be contained in the object
      */
-    public /*constructor*/ MyNumber(int v) {
+    public /*constructor*/ MyNumber(double v) {
 	  value=v;
 	  }
 
@@ -70,7 +70,7 @@ public class MyNumber implements Expression
      */
   @Override
   public String toString() {
-	  return Integer.toString(value);
+	  return Double.toString(value);
   }
 
   /** Two MyNumber expressions are equal if the values they contain are equal
@@ -105,7 +105,7 @@ public class MyNumber implements Expression
      */
   @Override
   public int hashCode() {
-		return value;
+		return (int)value;
   }
 
 }
