+++ src/main/java/calculator/Operation.java
@@ -27,7 +27,7 @@ public abstract class Operation implements Expression
   /**
    * The neutral element of the operation (e.g. 1 for *, 0 for +)
    */
-  protected int neutral;
+  protected double neutral;
 
 	/** It is not allowed to construct an operation with a null list of expressions.
 	 * Note that it is allowed to have an EMPTY list of arguments.
@@ -60,7 +60,7 @@ public abstract class Operation implements Expression
 	 * @param r	second argument of the binary operation
 	 * @return	result of computing the binary operation
 	 */
-   public abstract int op(int l, int r);
+   public abstract double op(double l, double r);
     // the operation itself is specified in the subclasses
 
 	/** Add more parameters to the existing list of parameters
@@ -120,7 +120,7 @@ public abstract class Operation implements Expression
 	public int hashCode()
 	{
 		int result = 5, prime = 31;
-		result = prime * result + neutral;
+		result = prime * result + (int) neutral;
 		result = prime * result + symbol.hashCode();
 		result = prime * result + args.hashCode();
 		return result;
