+++ src/main/java/visitor/Evaluator.java
@@ -17,13 +17,13 @@ public class Evaluator extends Visitor {
     public Evaluator() {}
 
     /** The result of the evaluation will be stored in this private variable */
-    private int computedValue;
+    private double computedValue;
 
     /** getter method to obtain the result of the evaluation
      *
-     * @return an Integer object containing the result of the evaluation
+     * @return an Double object containing the result of the evaluation
      */
-    public Integer getResult() { return computedValue; }
+    public double getResult() { return computedValue; }
 
     /** Use the visitor design pattern to visit a number.
      *
@@ -38,14 +38,14 @@ public class Evaluator extends Visitor {
      * @param o The operation being visited
      */
     public void visit(Operation o) {
-        ArrayList<Integer> evaluatedArgs = new ArrayList<>();
+        ArrayList<Double> evaluatedArgs = new ArrayList<>();
         //first loop to recursively evaluate each subexpression
         for(Expression a:o.args) {
             a.accept(this);
             evaluatedArgs.add(computedValue);
         }
         //second loop to accumulate all the evaluated subresults
-        int temp = evaluatedArgs.get(0);
+        double temp = evaluatedArgs.get(0);
         int max = evaluatedArgs.size();
         for(int counter=1; counter<max; counter++) {
             temp = o.op(temp,evaluatedArgs.get(counter));
